--- a/component_tests/cgv/package.json
+++ b/component_tests/cgv/package.json
@@ -10,6 +10,8 @@
     "@testing-library/jest-dom": "^5.16.4",
     "@testing-library/react": "^12.1.4",
     "@testing-library/user-event": "^13.5.0",
+    "@rsbuild/core": "^1.0.1-rc.2",
+    "@rsbuild/plugin-react": "^1.0.1-rc.2",
     "@types/jest": "^27.4.1",
     "@types/node": "^16.11.26",
     "@types/react": "^18.0.0",
@@ -17,7 +19,6 @@
     "cypress": "^12.9.0",
     "react": "^17.0.0",
     "react-dom": "^17.0.0",
-    "react-scripts": "^5.0.1",
     "serve": "^14.2.0",
     "start-server-and-test": "^2.0.0",
     "typescript": "^5.0.4",
@@ -47,17 +48,9 @@
     "@jbrowse/react-circular-genome-view": "file:./packed/jbrowse-react-circular-genome-view.tgz"
   },
   "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject",
-    "test:e2e": "start-test 'serve --listen 5000 build/' 5000 'cypress run --headless --browser chrome'"
-  },
-  "eslintConfig": {
-    "extends": [
-      "react-app",
-      "react-app/jest"
-    ]
+    "start": "rsbuild start",
+    "build": "rsbuild build",
+    "test:e2e": "start-test 'serve --listen 5000 dist/' 5000 'cypress run --headless --browser chrome'"
   },
   "browserslist": {
     "production": [