--- a/component_tests/lgv/package.json
+++ b/component_tests/lgv/package.json
@@ -3,13 +3,14 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@craco/craco": "^7.0.0",
     "@fontsource/roboto": "^4.5.5",
     "@jbrowse/react-linear-genome-view": "^2.0.0",
     "@testing-library/cypress": "^9.0.0",
     "@testing-library/jest-dom": "^5.16.4",
     "@testing-library/react": "^12.1.4",
     "@testing-library/user-event": "^13.5.0",
+    "@rsbuild/core": "^1.0.1-rc.2",
+    "@rsbuild/plugin-react": "^1.0.1-rc.2",
     "@types/jest": "^27.4.1",
     "@types/node": "^16.11.26",
     "@types/react": "^18.0.0",
@@ -46,17 +47,9 @@
     "@jbrowse/react-linear-genome-view": "file:./packed/jbrowse-react-linear-genome-view.tgz"
   },
   "scripts": {
-    "start": "craco start",
-    "build": "craco build",
-    "test": "craco test",
-    "eject": "craco eject",
-    "test:e2e": "start-test 'serve --listen 5000 build/' 5000 'cypress run --headless --browser chrome'"
-  },
-  "eslintConfig": {
-    "extends": [
-      "react-app",
-      "react-app/jest"
-    ]
+    "start": "rsbuild start",
+    "build": "rsbuild build",
+    "test:e2e": "start-test 'serve --listen 5000 dist/' 5000 'cypress run --headless --browser chrome'"
   },
   "browserslist": {
     "production": [