--- a/eslint.config.mjs
+++ b/eslint.config.mjs
@@ -14,6 +14,8 @@ export default tseslint.config(
       '**/esm/**/*',
       '**/public/**/*',
       '**/storybook-static/**',
+      '**/rspack.config.js',
+      '**/rsbuild.config.js',
       'website/*',
       'packages/core/util/nanoid.js',
       'products/**/webpack.config.js',