--- a/integration.test.js
+++ b/integration.test.js
@@ -2,12 +2,12 @@ const fs = require('fs')
 
 if (process.env.BUILT_TESTS) {
   test('JBrowse Web built', () => {
-    expect(fs.existsSync('products/jbrowse-web/build/index.html')).toBeTruthy()
+    expect(fs.existsSync('products/jbrowse-web/dist/index.html')).toBeTruthy()
   })
 
   test('JBrowse Desktop built', () => {
     expect(
-      fs.existsSync('products/jbrowse-desktop/build/index.html'),
+      fs.existsSync('products/jbrowse-desktop/dist/index.html'),
     ).toBeTruthy()
   })
 } else {