--- a/products/jbrowse-desktop/rsbuild.config.js
+++ b/products/jbrowse-desktop/rsbuild.config.js
@@ -1,0 +1,15 @@
+const { defineConfig } = require('@rsbuild/core')
+
+module.exports = defineConfig({
+  tools: {
+    rspack: {
+      target: 'electron-renderer',
+    },
+  },
+  html: {
+    title: 'JBrowse',
+    meta: {
+      description: 'A fast and flexible genome browser',
+    },
+  },
+})