--- a/products/jbrowse-react-circular-genome-view/package.json
+++ b/products/jbrowse-react-circular-genome-view/package.json
@@ -30,8 +30,8 @@
     "build:esm": "tsc -b tsconfig.build.esm.json",
     "build:commonjs": "tsc -b tsconfig.build.commonjs.json",
     "build:webpack": "npm-run-all build:webpack:*",
-    "build:webpack:dev": "cross-env NODE_ENV=development webpack",
-    "build:webpack:prod": "cross-env NODE_ENV=production webpack",
+    "build:webpack:dev": "cross-env NODE_ENV=development rspack build",
+    "build:webpack:prod": "cross-env NODE_ENV=production rspack build",
     "clean": "rimraf esm dist *.tsbuildinfo",
     "storybook": "storybook dev --port 6006",
     "storybook:build": "storybook build"