--- a/products/jbrowse-react-circular-genome-view/rspack.config.js
+++ b/products/jbrowse-react-circular-genome-view/rspack.config.js
@@ -1,5 +1,5 @@
+const rspack = require('@rspack/core')
 const path = require('path')
-const webpack = require('webpack')
 
 const buildDir = path.resolve('.')
 const distDir = path.resolve(buildDir, 'dist')
@@ -9,26 +9,27 @@ const mode = process.env.NODE_ENV || 'production'
 module.exports = {
   mode,
   entry: path.join(buildDir, 'src', 'index.ts'),
-  devtool: process.env.NODE_ENV === 'development' ? 'source-map' : false,
+  devtool: 'source-map',
   output: {
     path: distDir,
     filename:
       mode === 'production'
-        ? 'react-linear-genome-view.umd.production.min.js'
-        : 'react-linear-genome-view.umd.development.js',
-    library: 'JBrowseReactLinearGenomeView',
+        ? 'react-circular-genome-view.umd.production.min.js'
+        : 'react-circular-genome-view.umd.development.js',
+    library: 'JBrowseReactCircularGenomeView',
     libraryTarget: 'umd',
   },
   devServer: {
     port: 9000,
-    open: 'umd_example/',
+    open: true,
+    openPage: 'umd_example/',
     headers: {
       'Access-Control-Allow-Origin': '*',
       'Access-Control-Allow-Headers': '*',
     },
   },
   plugins: [
-    new webpack.optimize.LimitChunkCountPlugin({
+    new rspack.optimize.LimitChunkCountPlugin({
       maxChunks: 1,
     }),
   ],
@@ -51,19 +52,17 @@ module.exports = {
   module: {
     rules: [
       {
-        oneOf: [
-          {
-            test: /\.m?[tj]sx?$/,
-            exclude: /(node_modules|bower_components)/,
-            use: {
-              loader: 'babel-loader',
-              options: {
-                rootMode: 'upward',
-                presets: ['@babel/preset-react'],
-              },
+        test: /\.tsx?$/,
+        exclude: [/node_modules/],
+        loader: 'builtin:swc-loader',
+        options: {
+          jsc: {
+            parser: {
+              syntax: 'typescript',
             },
           },
-        ],
+        },
+        type: 'javascript/auto',
       },
     ],
   },