--- a/products/jbrowse-react-linear-genome-view/.storybook/main.js
+++ b/products/jbrowse-react-linear-genome-view/.storybook/main.js
@@ -2,28 +2,13 @@ module.exports = {
   stories: ['../stories/**/*.mdx', '../stories/**/*.stories.@(js|jsx|ts|tsx)'],
   staticDirs: ['../public'],
   addons: ['@storybook/addon-essentials'],
-  framework: {
-    name: '@storybook/react-webpack5',
-    options: {},
+  framework: 'storybook-react-rsbuild',
+  typescript: {
+    reactDocgen: false,
   },
-  docs: {
-    docsPage: true,
-  },
-
-  webpackFinal: async config => {
-    config.module.rules.push({
-      test: /\.(ts|tsx|js|jsx)$/,
-      use: [
-        {
-          loader: require.resolve('babel-loader'),
-          options: {
-            rootMode: 'upward',
-            presets: ['@babel/preset-react'],
-          },
-        },
-      ],
-    })
-    config.resolve.extensions.push('.ts', '.tsx')
+  rsbuildFinal: config => {
+    config.output ||= {}
+    config.output.assetPrefix = './'
     return config
   },
 }