--- a/products/jbrowse-react-linear-genome-view/rspack.config.js
+++ b/products/jbrowse-react-linear-genome-view/rspack.config.js
@@ -1,5 +1,5 @@
+const rspack = require('@rspack/core')
 const path = require('path')
-const webpack = require('webpack')
 
 const buildDir = path.resolve('.')
 const distDir = path.resolve(buildDir, 'dist')
@@ -9,27 +9,26 @@ const mode = process.env.NODE_ENV || 'production'
 module.exports = {
   mode,
   entry: path.join(buildDir, 'src', 'index.ts'),
-  devtool: process.env.NODE_ENV === 'development' ? 'source-map' : false,
+  devtool: 'source-map',
   output: {
     path: distDir,
     filename:
       mode === 'production'
-        ? 'react-circular-genome-view.umd.production.min.js'
-        : 'react-circular-genome-view.umd.development.js',
-    library: 'JBrowseReactCircularGenomeView',
+        ? 'react-linear-genome-view.umd.production.min.js'
+        : 'react-linear-genome-view.umd.development.js',
+    library: 'JBrowseReactLinearGenomeView',
     libraryTarget: 'umd',
   },
   devServer: {
     port: 9000,
-    open: true,
-    openPage: 'umd_example/',
+    open: 'umd_example/',
     headers: {
       'Access-Control-Allow-Origin': '*',
       'Access-Control-Allow-Headers': '*',
     },
   },
   plugins: [
-    new webpack.optimize.LimitChunkCountPlugin({
+    new rspack.optimize.LimitChunkCountPlugin({
       maxChunks: 1,
     }),
   ],
@@ -52,19 +51,17 @@ module.exports = {
   module: {
     rules: [
       {
-        oneOf: [
-          {
-            test: /\.m?[tj]sx?$/,
-            exclude: /(node_modules|bower_components)/,
-            use: {
-              loader: 'babel-loader',
-              options: {
-                rootMode: 'upward',
-                presets: ['@babel/preset-react'],
-              },
+        test: /\.tsx?$/,
+        exclude: [/node_modules/],
+        loader: 'builtin:swc-loader',
+        options: {
+          jsc: {
+            parser: {
+              syntax: 'typescript',
             },
           },
-        ],
+        },
+        type: 'javascript/auto',
       },
     ],
   },