--- a/products/jbrowse-web/package.json
+++ b/products/jbrowse-web/package.json
@@ -3,15 +3,14 @@
   "version": "2.15.0",
   "license": "Apache-2.0",
   "homepage": ".",
-  "main": "build",
   "dir": "src",
   "scripts": {
-    "start": "node scripts/start.js",
-    "build": "node scripts/build.js",
-    "postbuild": "node -e \"fs.writeFileSync('build/version.txt',JSON.parse(require('fs').readFileSync('package.json', 'utf8')).version)\"",
+    "build": "rsbuild build",
+    "dev": "tsx start.js",
+    "postbuild": "node -e \"fs.writeFileSync('dist/version.txt',JSON.parse(require('fs').readFileSync('package.json', 'utf8')).version)\"",
     "prepack": "yarn build",
-    "start-oauth-server": "ts-node --project tsconfig.oauth.json src/tests/OAuthServer/app.ts",
-    "start-basicauth-server": "ts-node --project tsconfig.oauth.json src/tests/HTTPBasicAuthServer/app.ts"
+    "start-oauth-server": "tsx src/tests/OAuthServer/app.ts",
+    "start-basicauth-server": "tsx src/tests/HTTPBasicAuthServer/app.ts"
   },
   "dependencies": {
     "@fontsource/roboto": "^5.0.2",
@@ -50,6 +49,8 @@
     "@mui/icons-material": "^6.0.0",
     "@mui/material": "^6.0.0",
     "@mui/x-data-grid": "^7.0.0",
+    "@rsbuild/core": "^1.0.1-rc.2",
+    "@rsbuild/plugin-react": "^1.0.1-rc.2",
     "base64-js": "^1.3.0",
     "clone": "^2.1.2",
     "copy-to-clipboard": "^3.3.1",
@@ -61,7 +62,6 @@
     "pako": "^2.1.0",
     "prompts": "^2.4.2",
     "react": "^18.0.0",
-    "react-dev-utils": "^12.0.1",
     "react-dom": "^18.0.0",
     "react-error-boundary": "^4.0.3",
     "requestidlecallback-polyfill": "^1.0.2",