--- a/products/jbrowse-web/rsbuild.config.js
+++ b/products/jbrowse-web/rsbuild.config.js
@@ -1,0 +1,14 @@
+const { defineConfig } = require('@rsbuild/core')
+
+module.exports = defineConfig({
+  output: {
+    assetPrefix: 'auto',
+  },
+  html: {
+    title: 'JBrowse',
+    favicon: 'public/favicon.ico',
+    meta: {
+      description: 'A fast and flexible genome browser',
+    },
+  },
+})