--- a/products/jbrowse-web/tsconfig.oauth.json
+++ b/products/jbrowse-web/tsconfig.oauth.json
@@ -1,6 +1,0 @@
-{
-  "include": ["src/tests/oauthServer"],
-  "compilerOptions": {
-    "module": "nodenext"
-  }
-}