--- a/webpack/config/webpack/persistentCache/createEnvironmentHash.js
+++ b/webpack/config/webpack/persistentCache/createEnvironmentHash.js
@@ -1,8 +1,0 @@
-const { createHash } = require('crypto')
-
-module.exports = function createEnvironmentHash(env) {
-  const hash = createHash('md5')
-  hash.update(JSON.stringify(env))
-
-  return hash.digest('hex')
-}