… component_tests/lgv/craco.config.js
… component_tests/react-app/craco.config.js
… package.json
… products/jbrowse-desktop/package.json
… products/jbrowse-desktop/scripts/config.js
… products/jbrowse-react-linear-genome-view/package.json
… products/jbrowse-web/public/index.html
… products/jbrowse-web/scripts/config.js
… webpack/config/env.js
… webpack/config/getHttpsConfig.js
… webpack/config/modules.js
… webpack/config/paths.js
… webpack/config/webpack.config.js
… webpack/scripts/build.js
… webpack/scripts/start.js
… yarn.lock