This PR proposes using rspack, a "webpack-compatible" js bundler, for dev and production builds of jbrowse2

The rsbuild ecosystem was relatively easy to swap out for our codebase

## why change at all

- reduced config
- substantial speedups

## why rspack?

- rspack's high compatibility with webpack configs made it very easy to switch
- it also allowed us to remove a lot of webpack config from the ejected cra config because it actually has many nice 'batteries included' features like typescript support, allowing us to remove a lot of babel config
- made dev server, production builds, storybook, and full repo ci much faster
- actually makes bundle size smaller (710kb -> 650kb for gff loading, 1.5Mb -> 942kb for alignments track)
- nice docs, reached "v1.0.0 stability" recently https://rspack.dev/


speed

| app    |before time | after time|speedup|
| -------- | ------- |------|----|
| jbrowse-web yarn start  | 19s    |1s|~19x|
| jbrowse-web yarn build | 27s     |3s|~9x|
| @jbrowse/react-app yarn storybook|100s|1s|~100x**|
| full repo CI|18minutes|10minutes|~1.8x|

**note: this included disabling `react-docgen` which was slow and seemed unused (?)


## why would we maybe not rspack/rsbuild

- still sort of new
- could be fast evolving
- less battle tested (however some components it uses like swc are quite battle tested being used in nextjs, etc.)
- could encounter bugs (i would imagine most bugs would be dev facing (e.g. build error) rather than user facing though)

## why not vite?

vite is arguably the most popular bundler, however we have trouble with it

- i have tried making it work but it runs into show stopper errors (see screenshot and branch below)
- even when it partially compiles, it is kinda slow
- gives a bunch of warnings at startup
- has trouble with our 'react on the webworker' construct

Screenshot of vite giving an error:
<details>

![Screenshot from 2024-09-03 00-01-51](https://github.com/user-attachments/assets/00049b92-f880-47c7-a313-06125adcb4bc)

</details>


vite branch of jbrowse-web https://github.com/GMOD/jbrowse-components/tree/vite_web2